requirements.txt
google-cloud-vision
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pyzmq==26.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
unicode==2.9
Unidecode==1.3.8
uvicorn==0.34.0
wcwidth==0.2.13
transformers==4.34.0
datasets==2.14.4
torch
tqdm==4.66.1
scikit-learn==1.3.0